Base user authorities and roles based on CMDB
The process of fetching owners of providers is placed in CmdbOwnerService.
It fetches info about all the CMDB providers and caches these values
in a manner, which makes querying it fast. The cache is invalidated
every x seconds.

The process of determining correct authorities is performed in PortalUserImpl
every call to the getAuthorities method, which guarantees the application
always gets correct authorities.